Abbreviate game counts past a billion, and sum totals as Long
Installing on the real watch showed the ECO chip reading "1927.8M": the
formatter stopped at millions, and the starting position filtered to the
signed-in account's own rating bands really does have 1.9 billion games.
No emulator fixture was ever that large.

That figure is also 90% of Int.MAX, so summing the three colours as Int
is one busier position away from wrapping negative — the totals are Long
now. The formatter moved to :core so both facts are pinned by tests.

+++ app/src/main/kotlin/com/closeonjae/chesspuzzle/opening/OpeningViewModel.kt
@@ -47,7 +47,7 @@ data class OpeningUiState(
     val isError: Boolean = false,
     val isSheetOpen: Boolean = false,
 ) {
-    val total: Int get() = white + draws + black
+    val total: Long get() = white.toLong() + draws + black
 }
 
 /**
